# internal/controller/codegen/httpapi/task_models.gen.go
// Package httpapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.12.4 DO NOT EDIT.
package httpapi

import (
	"time"
)

// Defines values for CreateTaskBodyStatus.
const (
	CreateTaskBodyStatusClosed CreateTaskBodyStatus = "closed"
	CreateTaskBodyStatusOpened CreateTaskBodyStatus = "opened"
)

// Defines values for TaskStatus.
const (
	TaskStatusClosed TaskStatus = "closed"
	TaskStatusOpened TaskStatus = "opened"
)

// CreateTaskBody defines model for CreateTaskBody.
type CreateTaskBody struct {
	Description *string              `json:"description,omitempty"`
	Name        string               `json:"name"`
	Status      CreateTaskBodyStatus `json:"status"`
}

// CreateTaskBodyStatus defines model for CreateTaskBody.Status.
type CreateTaskBodyStatus string

// Error defines model for Error.
type Error struct {
	Message *string `json:"message,omitempty"`
}

// FilterParameters defines model for FilterParameters.
type FilterParameters struct {
	CreatedAtFrom *time.Time `json:"createdAtFrom,omitempty"`
	CreatedAtTo   *time.Time `json:"createdAtTo,omitempty"`
	Name          *string    `json:"name,omitempty"`
	UpdatedAtFrom *time.Time `json:"updatedAtFrom,omitempty"`
	UpdatedAtTo   *time.Time `json:"updatedAtTo,omitempty"`
}

// Task defines model for Task.
type Task struct {
	CreatedAt   time.Time  `json:"createdAt"`
	Description string     `json:"description"`
	Id          string     `json:"id"`
	Name        string     `json:"name"`
	Status      TaskStatus `json:"status"`
	UpdateAt    *time.Time `json:"updateAt,omitempty"`
}

// TaskStatus defines model for Task.Status.
type TaskStatus string

// TaskList defines model for TaskList.
type TaskList struct {
	Amount   int    `json:"amount"`
	Entities []Task `json:"entities"`
}

// GetTaskListParams defines parameters for GetTaskList.
type GetTaskListParams struct {
	// FilterParameters The filter parameters for getting the list of tasks
	FilterParameters *FilterParameters `form:"filterParameters,omitempty" json:"filterParameters,omitempty"`
}

// CreateTaskJSONRequestBody defines body for CreateTask for application/json ContentType.
type CreateTaskJSONRequestBody = CreateTaskBody

// UpdateTaskJSONRequestBody defines body for UpdateTask for application/json ContentType.
type UpdateTaskJSONRequestBody = Task